Fixed Issue & Added Comments (Properly)
Fixed issue involving throwing out of a non-terminal parse in the middle
of an input line. If the non-terminal was optional, the parse for the
next possible non-terminal would start in the middle of the line,
ensuring failure.

Also added comments to code done by Robert to offer my suggestions.

Lastly, continued implementing Input class.