Input Class is Feature Complete!
Input Class is now in a state of functionality. ImportTest has been run
and verified to work correctly.

Testing will need to be done to see what the parser fails to correctly
read or succeeds in incorrectly reading.

Also, included some comments and requirements on the code by Robert from
when this brach split off. Some may no longer apply.

Changes will need to be made after the merge!